refactor: extract shared test utilities to eliminate DRY violations
Centralises duplicated helpers (mockSession, waitFor, createQueryClient,
mockApproval, renderer query helpers) into mobile/src/__test-utils__/
so each test file imports from a single source of truth.
